refactor: 拆分Practice Schemas
